+++ src/main/java/org/nrg/containers/services/impl/CommandResolutionServiceImpl.java
@@ -1017,7 +1017,7 @@ public class CommandResolutionServiceImpl implements CommandResolutionService {
             }
             for (final CommandWrapperDerivedInput input : commandWrapper.derivedInputs()) {
                 // Derived inputs must have a non-blank parent name
-                final String parentName = input.derivedFromXnatInput();
+                final String parentName = input.derivedFromWrapperInput();
                 if (StringUtils.isBlank(parentName)) {
                     // This is unlikely to happen. This should be caught by command validation.
                     final String message = String.format("Derived input \"%s\" needs a parent.", input);
@@ -1467,13 +1467,13 @@ public class CommandResolutionServiceImpl implements CommandResolutionService {
 
                 // Fail fast: if we will not be able to create the output, either throw or log that now and don't try later.
                 // First check that the handler input has a unique value
-                final ResolvedInputValue parentInputResolvedValue = getInputValueByName(commandOutputHandler.xnatInputName(), resolvedInputTrees);
+                final ResolvedInputValue parentInputResolvedValue = getInputValueByName(commandOutputHandler.wrapperInputName(), resolvedInputTrees);
                 if (parentInputResolvedValue == null) {
                     final String message = String.format("Cannot resolve output \"%s\". " +
                                     "Input \"%s\" is supposed to handle the output, but it does not have a uniquely resolved value. " +
                                     "Either there is no value, or there are multiple values." +
                                     "(We can't loop over input values yet, so the latter is an error as much as the former.)",
-                            commandOutput.name(), commandOutputHandler.xnatInputName());
+                            commandOutput.name(), commandOutputHandler.wrapperInputName());
                     if (Boolean.TRUE.equals(commandOutput.required())) {
                         throw new CommandResolutionException(message);
                     } else {
@@ -1495,7 +1495,7 @@ public class CommandResolutionServiceImpl implements CommandResolutionService {
                 if (uri == null || !(uri instanceof ArchiveItemURI)) {
                     final String message = String.format("Cannot resolve output \"%s\". " +
                                     "Input \"%s\" is supposed to handle the output, but it does not have an XNAT object value.",
-                            commandOutput.name(), commandOutputHandler.xnatInputName());
+                            commandOutput.name(), commandOutputHandler.wrapperInputName());
                     if (Boolean.TRUE.equals(commandOutput.required())) {
                         throw new CommandResolutionException(message);
                     } else {
@@ -1518,7 +1518,7 @@ public class CommandResolutionServiceImpl implements CommandResolutionService {
                     final String message = String.format("Cannot resolve output \"%s\". " +
                                     "Input \"%s\" is supposed to handle the output, but user \"%s\" does not have permission " +
                                     "to edit the XNAT object \"%s\".",
-                            commandOutput.name(), commandOutputHandler.xnatInputName(),
+                            commandOutput.name(), commandOutputHandler.wrapperInputName(),
                             userI.getLogin(), parentValue);
                     if (Boolean.TRUE.equals(commandOutput.required())) {
                         throw new CommandResolutionException(message);
@@ -1535,7 +1535,7 @@ public class CommandResolutionServiceImpl implements CommandResolutionService {
                         .mount(commandOutput.mount())
                         .glob(commandOutput.glob())
                         .type(commandOutputHandler.type())
-                        .handledByXnatCommandInput(commandOutputHandler.xnatInputName())
+                        .handledByWrapperInput(commandOutputHandler.wrapperInputName())
                         .path(resolveTemplate(commandOutput.path(), resolvedInputValuesByReplacementKey))
                         .label(resolveTemplate(commandOutputHandler.label(), resolvedInputValuesByReplacementKey))
                         .build();