+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Not yet released
 * Added Changelog
 * Check if command exists before saving, rather than catching exception when it fails to save.
 * [CS-205][] Add `/containers/{id}/logs` and `/containers/{id}/logs/{stdout|stderr}` APIs to fetch container logs.
+* Rename wrapper input properties to remove "xnat":
+    * `derived-from-xnat-input` -> `derived-from-wrapper-input`
+    * `as-a-child-of-xnat-input` -> `as-a-child-of-wrapper-input`
 
 [CS-205]: https://issues.xnat.org/browse/CS-205